Temp commit - Improved checkpoint loading logic in `play.py` and `ts_test.py` to handle different formats and added error handling for better robustness. - Adjusted the `main` function in `play.py` to merge command line arguments with existing configurations. - Enhanced the `depth_to_heightmap` function to ensure non-negative heightmap values and added visualization for debugging. - Updated container dimensions in the configuration to accommodate changes in packing behavior. - Indicated a dirty state in the `rsl_rl` submodule to reflect uncommitted changes.

+++ scripts/tianshou/envs/Packing/container.py
@@ -544,7 +544,12 @@ class Container(object):
         Returns:
             list: _description_
         """
-        heightmap = copy.deepcopy(self.heightmap)
+        heightmap = copy.deepcopy(self.heightmap.astype(np.int32))
+        # import matplotlib.pyplot as plt
+        # plt.imshow(heightmap)
+        # plt.colorbar()
+        # plt.savefig("heightmap.png")
+        # plt.close()
         # left-bottom & right-top pos [bx, by, bz, tx, ty, tz], 
         # dimension: [tx-bx, ty-by, tz-bz],
         all_ems = compute_ems(heightmap, container_h=self.dimension[2])
